Create scraper.py
This is the first version of the scraper. It is designed to work on any social media site. There is one known issue, which is that if the user does not have a reddID in their bio, but a reddID appears somewhere else on their page (like a retweet), there could be potential false positives.